fix: remove dtype kwarg from ascontiguousarray for numpy compatibility

+++ rag_engine.py
@@ -60,10 +60,10 @@ def retrieve(query: str, top_k: int = 3):
 
     index, corpus = load_index()
 
-    q_emb = model.encode([query], convert_to_numpy=True).astype(np.float32)
-    q_emb = q_emb / np.linalg.norm(q_emb, axis=1, keepdims=True)
-    # FAISS requires C-contiguous, strictly float32 arrays
-    q_emb = np.ascontiguousarray(q_emb, dtype=np.float32)
+    q_emb = model.encode([query], convert_to_numpy=True)
+    q_emb = (q_emb / np.linalg.norm(q_emb, axis=1, keepdims=True)).astype(np.float32)
+    # FAISS requires C-contiguous arrays
+    q_emb = np.ascontiguousarray(q_emb)
 
     scores, indices = index.search(q_emb, top_k)
